subject

+++ StudentLibrary/RepositoryEF/IStudentRepositoryEF.cs
@@ -9,7 +9,7 @@ namespace EFDataAccessLayer.RepositoryEF
     public  interface IStudentRepositoryEF
     {
         public IEnumerable<StudentDetails> GetAllDetails();
-        public StudentDetails GetbyID(long studentid);
+        public StudentDetails GetbyID(int studentid);
         public void Insert(StudentDetails stud);
         public void Update(int id, StudentDetails stud);
         public void Delete(int studentId);
